Changes user controller to be a singleton
The controller model used in the rest of the code is to use singletons.
So instead of doing new CurrentUserController().getUser(), you do
CurrentUserController.getInstance().getUser(). This makes sense because
there should only be one instance of the controller at a time.
Also removed tests for methods that were deleted in GameModel.
Also now recognizes the user who is logged in and won't let them
estimate a requirement twice.