Adds basic updating of requirements in Requirement Manager
The estimate for a requirement in Requirement Manager can now be updated. GetParentRequirementController needs to be threaded like the different UserController classes so that the main thread is blocked until a list of requirements is received. Right now, it requires the user to press the send to requirement manager button twice since the first time it requests the requirements, but does not receive them before it attempts to update the actual requirement.

Author: nfbrown