Adds a requirement request thread and properly updates requirements on the server
The tests will now run somewhat slowly because anything dependent on receiving requirements from the server now has to timeout before the tests can continue. The tests still work though. The only issue left to solve is updating the requirement manager to reflect the changes. Right now the changes are not reflected until Janeway is restarted.

Author: nfbrown